requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.9.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.1
pycodestyle==2.10.0
pycryptodomex==3.17
python-dotenv==1.0.0
regex==2022.10.31
requests==2.28.2
tiktoken==0.3.0
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
